WIP on cockney filter
TODO: I had misinterpreted the code. It's not a series
of find/replaces like LOLCAT and chef, but a lexer,
which means that once one of the find/replaces has worked
on a bit of code, none of the others will apply.

So I either need to implement it as a lexer myself, or
see if I can simplify it into a series of find/replaces.

It also turns out that, although the original version of this
is idempotent, that's because it seeds the random number
generator with a 0, which is not a behavior I can directly
replicate in JS (Math.random() doesn't even have a seed
option!). So there's no way this test will ever pass. But it's
still useful for visual comparison.