Use tiny PRNG instead of replacing rand() with modulo
Some of the original filters use random numbers in their
algorithms, which makes it non-trivial to test whether my
port creates "the same" output.

Previously I was dealing with this by replacing the random
numbers with counts and modulo's. But that gives output
that is a little bit too predictable. So instead, I'm using
a tiny pseudo-random number generator algorithm (specifically,
"MINSTD"), which I can implement identically in C, Perl, and JS.